Rename: 테스트 빌더의 이름을 프로덕션 코드의 네이밍과 통일시킴
